feat: add target Y coordinate and progress data to Quantum Quarry

+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/menu/quantum_quarry/QuantumQuarryMenu.kt
@@ -62,7 +62,7 @@ class QuantumQuarryMenu(
 	fun getTargetY(): Int = quarryContainerData.get(QuantumQuarryBlockEntity.TARGET_Y_DATA_INDEX)
 	fun getTargetZ(): Int = quarryContainerData.get(QuantumQuarryBlockEntity.TARGET_Z_DATA_INDEX)
 	fun getProgress(): Int = quarryContainerData.get(QuantumQuarryBlockEntity.PROGRESS_DATA_INDEX)
-	fun getAmountBlocksBroken(): Int = quarryContainerData.get(QuantumQuarryBlockEntity.AMOUNT_BLOCKS_BROKEN_DATA_INDEX)
+	fun getAmountBlocksBroken(): UInt = quarryContainerData.get(QuantumQuarryBlockEntity.AMOUNT_BLOCKS_BROKEN_DATA_INDEX).toUInt()
 	fun getBiomeId(): Int = quarryContainerData.get(QuantumQuarryBlockEntity.BIOME_ID_DATA_INDEX)
 
 	override fun quickMoveStack(player: Player, index: Int): ItemStack {
